requirements.txt
asgiref==3.7.2
Babel==2.13.0
Django==4.2.6
django-etc==1.4.0
django-hitcount==1.3.5
django-modeltranslation==0.18.11
Pillow==10.0.1
python-decouple==3.8
python-gettext==5.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
